+++ tests/targets/nester.py
@@ -6,9 +6,9 @@ Used to test nested modules.
 """
 from ec.ec import member, task
 
-import target_script
+import simple
 
-member(target_script) # add the imported member to the script
+member(simple) # add the imported member to the script
 
 @task
 def task1():