+++ tests/test_configuration.py
@@ -8,9 +8,9 @@ import unittest
 
 from ec import interface
 
-from support import target_script
+from targets import simple
 
-interface.setBase(target_script)
+interface.setBase(simple)
 
 class TestConfiguration(unittest.TestCase):
   def setUp(self):
@@ -20,7 +20,7 @@ class TestConfiguration(unittest.TestCase):
     pass
 
   def test_verify_configuration(self):
-    Members =  target_script.__ec_member__.Members
+    Members =  simple.__ec_member__.Members
     
     assert(set(['task1', 't1', 'group1', 'ex', 'hex']) == set(Members.keys()))
     assert(set(['task1']) == set(Members['group1'].Members.keys()))