+++ tests/test_shell.py
@@ -10,7 +10,7 @@ from support.helpers import shell_exec, checkResult
 def launch_ec(*lines):
   """Passes commands through the shell.
   """
-  return shell_exec('python tests/support/target_script.py', input='%s\n\0' % '\r\n'.join(lines))
+  return shell_exec('python tests/targets/simple.py', input='%s\n\0' % '\r\n'.join(lines))
 
 class TestShell(unittest.TestCase):
   def setUp(self):